app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.60.0
pydantic>=2.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
